requirements.txt
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
coverage==5.2.1
crayons==0.4.0
cryptography==35.0.0
cssselect==1.1.0
daphne==3.0.2
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
html5print==0.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.1
incremental==21.3.0
lxml==4.6.3
Markdown==3.3.4
markdown2==2.4.1
numpy==1.21.2
opencv-contrib-python==4.5.3.56
outcome==1.1.0
pandas==1.3.4
Pillow==8.3.2
platformdirs==2.3.0
ply==3.4
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2020.1
readtime==1.1.1
requests==2.26.0
selenium==4.0.0
service-identity==21.1.0
six==1.16.0
slimit==0.8.1
sniffio==1.2.0
sorl-thumbnail==12.7.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.3.1
tinycss2==1.1.0
toml==0.10.1
trio==0.19.0
trio-websocket==0.9.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.7.2
webdriver-manager==3.5.1
webencodings==0.5.1
whitenoise==5.3.0
wsproto==1.0.0
zipp==3.5.0
zope.interface==5.4.0
